# search: Improve performance of zoektIndexedRepos